# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Account:
    properties:
      balance:
        type: number
      created_at:
        type: string
      id:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
  models.AccountBalanceResponse:
    properties:
      account_id:
        type: string
      balance:
        type: number
      currency:
        type: string
    type: object
...
  models.CreateAccountRequest:
    properties:
      phone_number:
        type: string
    required:
    - phone_number
    type: object
...
  models.Discount:
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: invalid_credentials
        type: string
    type: object
...
  models.LoyaltyAccount:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      points:
        type: integer
      tier:
        type: string
      updated_at:
        type: string
    type: object
...
  models.TokenPairResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Comprehensive restaurant management system with ordering, payments,
    inventory, and analytics
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management System API
  version: "2.0"
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    get:
      description: Get account information by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get account details
      tags:
      - enterprise
    put:
      consumes:
      - application/json
      description: Update account information
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account update
        in: body
        name: request
        required: true
        schema:
          properties:
            name:
              type: string
            phone:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: Updated
      security:
      - BearerAuth: []
      summary: Update account
      tags:
      - enterprise
  /accounts/{id}/balance:
    get:
      description: Retrieve account balance by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountBalanceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get account balance
      tags:
      - accounts
  /accounts/{id}/loyalty:
    get:
      description: Get loyalty points for an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoyaltyAccount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get loyalty account
      tags:
      - enterprise
  /accounts/{id}/loyalty/earn:
    post:
      consumes:
      - application/json
      description: Add loyalty points to an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Points to earn
        in: body
        name: request
        required: true
        schema:
          properties:
            points:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: Points earned
      security:
      - BearerAuth: []
      summary: Earn loyalty points
      tags:
      - enterprise
  /accounts/{id}/roles:
    post:
      consumes:
      - application/json
      description: Assign a role to a user account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Role assignment
        in: body
        name: request
        required: true
        schema:
          properties:
            restaurant_id:
              type: string
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: Role assigned
      security:
      - BearerAuth: []
      summary: Assign role to account
      tags:
      - enterprise
  /accounts/{id}/roles/{role}:
    delete:
      description: Remove a role from a user account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Role name
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Role removed
      security:
      - BearerAuth: []
      summary: Remove role from account
      tags:
      - enterprise
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate refresh token and logout
      parameters:
      - description: Logout request
        in: body
        name: request
        required: true
        schema:
          properties:
            device_id:
              type: string
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenPairResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access and refresh tokens using refresh token
      parameters:
      - description: Refresh request
        in: body
        name: request
        required: true
        schema:
          properties:
            device_id:
              type: string
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenPairResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/request-otp:
    post:
      consumes:
      - application/json
      description: Send OTP to user's phone number
      parameters:
      - description: OTP request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Request OTP for authentication
      tags:
      - auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate with phone and password to receive a JWT token
      parameters:
      - description: Signin request
        in: body
        name: request
        required: true
        schema:
          properties:
            password:
              type: string
            phone:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Returns token, account_id, and role
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Sign in to an existing account
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account with phone and password
      parameters:
      - description: Signup request
        in: body
        name: request
        required: true
        schema:
          properties:
            password:
              type: string
            phone:
              type: string
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Sign up a new user
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verify OTP and receive access and refresh tokens
      parameters:
      - description: OTP verification
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Verify OTP code
      tags:
      - auth
  /branches/{branchId}/waitlist:
    post:
      consumes:
      - application/json
      description: Add customer to restaurant waitlist
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Join waitlist
      tags:
      - customer
  /branches/{branchId}/waitlist/{entryId}:
    patch:
      consumes:
      - application/json
      description: Update customer's waitlist entry
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: string
      - description: Entry ID
        in: path
        name: entryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update waitlist entry
      tags:
      - customer
  /categories:
    get:
      description: Get all menu categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new menu category
      parameters:
      - description: Category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Delete a menu category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Deleted
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update menu category details
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
  /discounts:
    post:
      consumes:
      - application/json
      description: Create a new discount code
      parameters:
      - description: Discount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Discount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Discount'
      security:
      - BearerAuth: []
      summary: Create discount
      tags:
      - enterprise
  /discounts/apply:
    post:
      consumes:
      - application/json
      description: Apply a discount code to an order
      parameters:
      - description: Discount application
        in: body
        name: request
        required: true
        schema:
          properties:
            account_id:
              type: string
            code:
              type: string
            order_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Apply discount code
      tags:
      - enterprise
  /favorites:
    post:
      consumes:
      - application/json
      description: Add a menu item to user's favorites
      parameters:
      - description: Favorite request
        in: body
        name: request
        required: true
        schema:
          properties:
            menu_item_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Favorite'
      security:
      - BearerAuth: []
      summary: Add favorite menu item
      tags:
      - favorites
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"